cosotas

+++ src/RenderSystem.cpp
@@ -33,11 +33,11 @@ namespace PitudoEngine {
 		#ifdef _DEBUG
 			if (collider.m_colliderShape == ColliderShape::RECT) {
 				Vec2 collPos = transform.position - (collider.m_Size * collider.m_pivot);
-				tigrFillRect(m_Window, collPos.x, collPos.y, collider.m_Size.x, collider.m_Size.y, tigrRGBA(0x00, 0xFF, 0x00, 0x80));
+				tigrFillRect(m_Window, (int)collPos.x, (int)collPos.y, (int)collider.m_Size.x, (int)collider.m_Size.y, tigrRGBA(0x00, 0xFF, 0x00, 0x80));
 			}
 			else {
 				Vec2 collPos = transform.position;
-				tigrFillCircle(m_Window, collPos.x, collPos.y, collider.m_Size.x, tigrRGBA(0x00, 0xFF, 0x00, 0x80));
+				tigrFillCircle(m_Window, (int)collPos.x, (int)collPos.y, (int)collider.m_Size.x, tigrRGBA(0x00, 0xFF, 0x00, 0x80));
 			}
 		#endif
